lr: Add a reference to StateMap to ParsingTable
