+++ parser/lr/canonical/parsing_table_test.go
@@ -9,101 +9,6 @@ import (
 	"github.com/moorara/algo/parser/lr"
 )
 
-var prods = [][]*grammar.Production{
-	{
-		{Head: "S′", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("S")}},                          // S′ → S
-		{Head: "S", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("C"), grammar.NonTerminal("C")}}, // S → CC
-		{Head: "C", Body: grammar.String[grammar.Symbol]{grammar.Terminal("c"), grammar.NonTerminal("C")}},    // C → cC
-		{Head: "C", Body: grammar.String[grammar.Symbol]{grammar.Terminal("d")}},                              // C → d
-	},
-	{
-		{Head: "E′", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("E")}},                                                 // E′ → E
-		{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("E"), grammar.Terminal("+"), grammar.NonTerminal("E")}}, // E → E + E
-		{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("E"), grammar.Terminal("*"), grammar.NonTerminal("E")}}, // E → E * E
-		{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.Terminal("("), grammar.NonTerminal("E"), grammar.Terminal(")")}},    // E → ( E )
-		{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.Terminal("id")}},                                                    // E → id
-	},
-	{
-		{Head: "grammar′", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("grammar")}},                           // grammar′ → grammar
-		{Head: "grammar", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("name"), grammar.NonTerminal("decls")}}, // grammar → name decls
-		{Head: "name", Body: grammar.String[grammar.Symbol]{grammar.Terminal("grammar"), grammar.Terminal("IDENT")}},       // name → "grammar" IDENT
-		{Head: "decls", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("decls"), grammar.NonTerminal("decl")}},   // decls → decls decl
-		{Head: "decls", Body: grammar.E}, // decls → ε
-		{Head: "decl", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("token")}},                                                  // decl → token
-		{Head: "decl", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rule")}},                                                   // decl → rule
-		{Head: "token", Body: grammar.String[grammar.Symbol]{grammar.Terminal("TOKEN"), grammar.Terminal("="), grammar.Terminal("STRING")}}, // token → TOKEN "=" STRING
-		{Head: "token", Body: grammar.String[grammar.Symbol]{grammar.Terminal("TOKEN"), grammar.Terminal("="), grammar.Terminal("REGEX")}},  // token → TOKEN "=" REGEX
-		{Head: "rule", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("lhs"), grammar.Terminal("="), grammar.NonTerminal("rhs")}}, // rule → lhs "=" rhs
-		{Head: "rule", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("lhs"), grammar.Terminal("=")}},                             // rule → lhs "="
-		{Head: "lhs", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("nonterm")}},                                                 // lhs → nonterm
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.NonTerminal("rhs")}},                         // rhs → rhs rhs
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("rhs"), grammar.Terminal("|"), grammar.NonTerminal("rhs")}},  // rhs → rhs "|" rhs
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.Terminal("("), grammar.NonTerminal("rhs"), grammar.Terminal(")")}},       // rhs → "(" rhs ")"
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.Terminal("["), grammar.NonTerminal("rhs"), grammar.Terminal("]")}},       // rhs → "[" rhs "]"
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.Terminal("{"), grammar.NonTerminal("rhs"), grammar.Terminal("}")}},       // rhs → "{" rhs "}"
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.Terminal("{{"), grammar.NonTerminal("rhs"), grammar.Terminal("}}")}},     // rhs → "{{" rhs "}}"
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("nonterm")}},                                                 // rhs → nonterm
-		{Head: "rhs", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("term")}},                                                    // rhs → term
-		{Head: "nonterm", Body: grammar.String[grammar.Symbol]{grammar.Terminal("IDENT")}},                                                  // nonterm → IDENT
-		{Head: "term", Body: grammar.String[grammar.Symbol]{grammar.Terminal("TOKEN")}},                                                     // term → TOKEN
-		{Head: "term", Body: grammar.String[grammar.Symbol]{grammar.Terminal("STRING")}},                                                    // term → STRING
-	},
-}
-
-var grammars = []*grammar.CFG{
-	grammar.NewCFG(
-		[]grammar.Terminal{"c", "d"},
-		[]grammar.NonTerminal{"S", "C"},
-		prods[0][1:],
-		"S",
-	),
-	grammar.NewCFG(
-		[]grammar.Terminal{"+", "*", "(", ")", "id"},
-		[]grammar.NonTerminal{"E"},
-		prods[1][1:],
-		"E",
-	),
-	grammar.NewCFG(
-		[]grammar.Terminal{"=", "|", "(", ")", "[", "]", "{", "}", "{{", "}}", "grammar", "IDENT", "TOKEN", "STRING", "REGEX"},
-		[]grammar.NonTerminal{"grammar", "name", "decls", "decl", "token", "rule", "lhs", "rhs", "nonterm", "term"},
-		prods[2][1:],
-		"grammar",
-	),
-}
-
-func getTestParsingTables() []*lr.ParsingTable {
-	pt0 := lr.NewParsingTable(
-		[]lr.State{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
-		[]grammar.Terminal{"c", "d", grammar.Endmarker},
-		[]grammar.NonTerminal{"S", "C"},
-	)
-
-	pt0.AddACTION(0, "c", &lr.Action{Type: lr.SHIFT, State: 5})
-	pt0.AddACTION(0, "d", &lr.Action{Type: lr.SHIFT, State: 7})
-	pt0.AddACTION(1, grammar.Endmarker, &lr.Action{Type: lr.ACCEPT})
-	pt0.AddACTION(9, "c", &lr.Action{Type: lr.SHIFT, State: 6})
-	pt0.AddACTION(9, "d", &lr.Action{Type: lr.SHIFT, State: 8})
-	pt0.AddACTION(5, "c", &lr.Action{Type: lr.SHIFT, State: 5})
-	pt0.AddACTION(5, "d", &lr.Action{Type: lr.SHIFT, State: 7})
-	pt0.AddACTION(7, "c", &lr.Action{Type: lr.REDUCE, Production: prods[0][3]})
-	pt0.AddACTION(7, "d", &lr.Action{Type: lr.REDUCE, Production: prods[0][3]})
-	pt0.AddACTION(4, grammar.Endmarker, &lr.Action{Type: lr.REDUCE, Production: prods[0][1]})
-	pt0.AddACTION(6, "c", &lr.Action{Type: lr.SHIFT, State: 6})
-	pt0.AddACTION(6, "d", &lr.Action{Type: lr.SHIFT, State: 8})
-	pt0.AddACTION(8, grammar.Endmarker, &lr.Action{Type: lr.REDUCE, Production: prods[0][3]})
-	pt0.AddACTION(2, "c", &lr.Action{Type: lr.REDUCE, Production: prods[0][2]})
-	pt0.AddACTION(2, "d", &lr.Action{Type: lr.REDUCE, Production: prods[0][2]})
-	pt0.AddACTION(3, grammar.Endmarker, &lr.Action{Type: lr.REDUCE, Production: prods[0][2]})
-
-	pt0.SetGOTO(0, "S", 1)
-	pt0.SetGOTO(0, "C", 9)
-	pt0.SetGOTO(9, "C", 4)
-	pt0.SetGOTO(5, "C", 2)
-	pt0.SetGOTO(6, "C", 3)
-
-	return []*lr.ParsingTable{pt0}
-}
-
 func TestBuildParsingTable(t *testing.T) {
 	pt := getTestParsingTables()
